fixup! Rename ws migrate to migrate-definition
